requirements.txt
telethon
python-dotenv
python-docx
psycopg2-binary
pytz
pyyaml
